LayoutTests/platform/ios-sim/fast/dom/Window/window-xy-properties-expected.txt

PASS event.clientX is 50
PASS event.clientY is 50
PASS event.pageX is 50
PASS event.pageY is 50
PASS event.layerX is 50
PASS event.layerY is 50
PASS event.offsetX is 42
PASS event.offsetY is 42
PASS event.clientX is 50
PASS event.clientY is 50
PASS event.pageX is 50
PASS event.pageY is 550
PASS event.layerX is 50
PASS event.layerY is 550
PASS event.offsetX is 42
PASS event.offsetY is 542
PASS event.clientX is 50
PASS event.clientY is 50
PASS event.pageX is 550
PASS event.pageY is 50
PASS event.layerX is 550
PASS event.layerY is 50
PASS event.offsetX is 542
PASS event.offsetY is 42
PASS ev.clientX is 50
PASS ev.clientY is 50
FAIL ev.pageX should be 550. Was 50.
PASS ev.pageY is 550
FAIL ev.layerX should be 550. Was 50.
PASS ev.layerY is 550
FAIL ev.offsetX should be 542. Was 42.
PASS ev.offsetY is 542
